requirements.txt
google-cloud-bigquery == 2.32.0
pandas == 1.4.0
pyarrow == 6.0.1
requests == 2.27.1